Refractor of TicketBuy
This is the start of refractoring TicketBuy.
